Added more behaviour controls through env vars

+++ src/rekordbox2plex/actions/PlaylistSync.py
@@ -8,8 +8,9 @@ from ..plex.data_types import Track, PlexPlaylist, PlexPlaylists
 from ..mappers.TrackIdMapper import TrackIdMapper
 from ..utils.progress_bar import progress_instance
 from ..utils.logger import logger
-from typing import List, Literal
+from ..utils.helpers import get_boolenv
 from ._ActionBase import ActionBase
+from typing import List, Literal
 
 
 class PlaylistSync(ActionBase):
@@ -45,7 +46,7 @@ class PlaylistSync(ActionBase):
                     if plex_playlist:
                         playlist_was_synced = self.sync_playlist_tracks(
                             rb_playlist, plex_playlist
-                        )  # Make sure tracks are up to date
+                        ) # Make sure tracks are up to date
                         if playlist_was_synced:
                             progress.update(
                                 task,
@@ -76,10 +77,11 @@ class PlaylistSync(ActionBase):
                     task,
                     description="[bold green]✔ Done! Rekordbox playlists are synchronized with Plex!",
                 )
-        self.delete_orphaned_playlists(rb_playlists or [], plex_playlists)
         logger.info(
             f"[bold green]✔ Result: {self.deleted} deleted, {self.updated} updated, and {self.created} created."
         )
+        if get_boolenv("DELETE_ORPHANED_PLAYLISTS", "false"):
+            self.delete_orphaned_playlists(rb_playlists or [], plex_playlists)
         logger.info(
             "[bold green]✔ Process complete! Rekordbox and Plex playlists should now be in sync!"
         )
